requirements.txt
python-telegram-bot==20.8
APScheduler==3.8.1
python-dotenv==1.0.0
pytz==2024.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
aiohttp==3.9.3         # 用于异步HTTP请求
beautifulsoup4==4.12.3 # 用于解析网页内容（可选，如果需要抓取链接标题）
lxml==5.1.0   